+++ bin/rialto/runtime/src/rialto_poa.rs
@@ -115,7 +115,7 @@ impl TPruningStrategy for PruningStrategy {
 pub struct ChainTime;
 
 impl TChainTime for ChainTime {
-	fn header_is_ahead(&self, timestamp: u64) -> bool {
+	fn is_timestamp_ahead(&self, timestamp: u64) -> bool {
 		let now = super::Timestamp::now();
 		timestamp > now
 	}
